[GOB-1027] Refactor folder structure
To prepare for the new types of importers.
